Throw exception when no stripe id is found
In cases where methods are called which expect the billable entity to be a Stripe customer we'll first do a check from now on to see if that's the case. Otherwise we'll throw an exception to notify the developer that the billable entity must first be registered as a Stripe customer.